fix(ui): fix 5 ESLint errors and remove || true mask from CI
- Remove `|| true` from ESLint CI step that silently swallowed all errors
- Extract Router component to src/Router.tsx to fix react-refresh/only-export-components
- Simplify main.tsx to only bootstrap and render
- Add connectRef in useSSE.ts to fix TDZ circular reference (react-hooks/exhaustive-deps)
- Use direct conditional returns in Router.tsx instead of variable assignment to fix react-hooks/static-components
- Replace ternary side-effect in StatusPage.tsx toggleCommit with if/else
- Move eslint-disable-next-line to correct line in StatusPage.tsx useEffect
